Add Firebase configuration and improve JWT handling; enhance order notification and pagination UI

+++ franchise-supply-platform/config/auth.php
@@ -129,8 +129,8 @@ return [
 
     'jwt' => [
         'secret' => env('JWT_SECRET'),
-        'ttl' => env('JWT_TTL', 60), // Time to live in minutes
-        'refresh_ttl' => env('JWT_REFRESH_TTL', 20160), // 2 weeks
+        'ttl' => (int) env('JWT_TTL', 60), // Time to live in minutes
+        'refresh_ttl' => (int) env('JWT_REFRESH_TTL', 20160), // 2 weeks
     ],
 
 ];
